Add support for Lutron Scenes
Scenes are named after the button's engraving name.

With the current basic implementation, it's possible some
false-positives are added to the configuration, which have
no associated actions (e.g., keypad had a shell of a "Button 3"
created, but not configured).  Recommend to explicitly
define what scenes are added on the views using the group.yaml
file to ensure it is not crowded with non-functional scenes.