Add support for scene activation events
Add an explicit event when a scene is activated, ignoring the
scene state.  This enables the event to be raised multiple
times even if the scene is already activated (e.g., pressing
all off multiple times continues to raise the event), which
could be useful if a device or service linked as an action
to the event is mis-behaving and does not properly receive
the event on the first try.